refactor(frontend): Helper method to get assignment names for courses
