+++ frontend/src/app/assignment/services/course.service.ts
@@ -9,6 +9,7 @@ import {environment} from '../../../environments/environment';
 import {UserService} from '../../user/user.service';
 import Course, {CourseAssignee, CourseStudent, CreateCourseDto, UpdateCourseDto} from '../model/course';
 import {StorageService} from "../../services/storage.service";
+import {ReadAssignmentDto} from "../model/assignment";
 
 @Injectable()
 export class CourseService {
@@ -17,6 +18,21 @@ export class CourseService {
     private userService: UserService,
     private storageService: StorageService,
   ) {
+  }
+
+  getAssignmentNames(assignments: (ReadAssignmentDto | undefined)[]): string[] {
+    if (!assignments.length) {
+      return [];
+    }
+    const firstTitle = assignments.find(a => a?.title)?.title ?? '';
+    if (assignments.length === 1) {
+      return [firstTitle];
+    }
+    let prefixLength = 0;
+    while (prefixLength < firstTitle.length && assignments.every(a => !a || a.title[prefixLength] === firstTitle[prefixLength])) {
+      prefixLength++;
+    }
+    return assignments.map(a => a ? a.title.slice(prefixLength) : '<deleted>');
   }
 
   // --------------- Draft ---------------
